requirements.txt
django
djangorestframework
drf-yasg
psycopg2
